+++ services/web/server/src/simcore_service_webserver/projects/plugin.py
@@ -18,7 +18,12 @@ from servicelib.aiohttp.rest_routing import (
 
 from .._constants import APP_OPENAPI_SPECS_KEY, APP_SETTINGS_KEY
 from .._resources import resources
-from . import projects_handlers, projects_nodes_handlers, projects_tags_handlers
+from . import (
+    projects_handlers,
+    projects_handlers_crud,
+    projects_nodes_handlers,
+    projects_tags_handlers,
+)
 from .projects_access import setup_projects_access
 from .projects_db import setup_projects_db
 
@@ -77,6 +82,7 @@ def setup_projects(app: web.Application) -> bool:
         _create_routes(
             "project",
             specs,
+            projects_handlers_crud,
             projects_handlers,
             projects_nodes_handlers,
             projects_tags_handlers,
